v2 - Creating projectiles and items
Week 7
TODO:
Fix Koopa's shell collision

... I hate collision detection

+++ SMB3_v2_C++/assets/classes/Game.cpp
@@ -220,7 +220,7 @@ void Game::_ParseSettings(std::string line) {
 	}
 
 	if (tokens.at(0) == "default") {
-		_defaultSceneID = std::stoi(tokens.at(1));
+		_defaultSceneID = std::stoul(tokens.at(1));
 	}
 	else {
 		OutputDebugStringA("[GAME] Unknown game settings\n");
